[webkitbot] Test suite cannot run due to undeclared https-proxy-agent dependency https://bugs.webkit.org/show_bug.cgi?id=322279 rdar://problem/185518939
Reviewed by Aakash Jain.

Running tests/WebKitBot.test.mjs locally fails because src/WebKitBot.mjs imports https-proxy-agent,
which is not declared in package.json. This means that`npm ci` alone resolves only the transitive
v5, and `import { HttpsProxyAgent }` fails against it because v5 is CommonJS with no statically
detectable named export. The deployed bot is unaffected, since the image supplies v8 and Node
resolves it natively.

Along with declaring the dependency, move the pure command parsing out of WebKitBot.mjs into
a new module with no imports at all, so the tests exercise it without loading the Slack clients
or the proxy agent. Other than a drive-by fix to drop an unnecessary escape in the revision
regexp (\@ -> @), this is a straight move.

* Tools/WebKitBot/package.json: Declare https-proxy-agent so `npm ci` matches the deployed image.
* Tools/WebKitBot/package-lock.json: Hoist https-proxy-agent v8, v5 stays nested under axios and jsdom.
* Tools/WebKitBot/src/CommandParser.mjs: Added. Moved from WebKitBot.mjs.
(parseBugId): Moved.
(parsePRUrl): Moved.
(extractRevision): Moved.
(buildGitWebkitRevertCommand): Moved.
(extractRevisionsAndReason): Moved.
(extractCommandAndArgs): Moved.
(extractTextIfMentioned): Moved.
* Tools/WebKitBot/src/WebKitBot.mjs: Import the parsing helpers from CommandParser.mjs.
* Tools/WebKitBot/tests/WebKitBot.test.mjs: Import from CommandParser.mjs.

Canonical link: https://commits.webkit.org/319804@main
