+++ Tools/WebKitBot/package.json
@@ -10,6 +10,7 @@
     "axios": "^1.15.1",
     "dotenv": "^8.2.0",
     "eslint": "^7.3.1",
+    "https-proxy-agent": "^8.0.0",
     "jest": "^26.1.0",
     "node-persist": "^3.1.0",
     "replaceall": "^0.1.6",
